backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
requests
httpx
aiohttp
beautifulsoup4
GitPython
python-owasp-zap-v2.4
python-jose
